requirements.txt
PyYAML>=6.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
requests>=2.25.0
pathlib2>=2.3.0
